Add permission for all pages

+++ app/Http/Controllers/GradeController.php
@@ -8,8 +8,17 @@ use Illuminate\Support\Facades\DB;
 
 class GradeController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index($slug)
     {
+        if (auth()->user()->role != 'student') {
+            return redirect()->back();
+        }
+
         $grades = DB::table('grades')
             ->join('users', 'grades.teacher', '=', 'users.id')
             ->where('student', '=', $slug)
@@ -20,6 +29,10 @@ class GradeController extends Controller
 
     public function show()
     {
+        if (auth()->user()->role == 'student') {
+            return redirect()->back();
+        }
+
         $grades = DB::table('grades')
             ->get();
 
@@ -28,6 +41,10 @@ class GradeController extends Controller
 
     public function details($slug)
     {
+        if (auth()->user()->role == 'student') {
+            return redirect()->back();
+        }
+
         $grade = Grade::find($slug);
 
         return view('grade-details', compact('grade'));
@@ -35,10 +52,23 @@ class GradeController extends Controller
 
     public function update($slug)
     {
+        if (auth()->user()->role == 'student') {
+            return redirect()->back();
+        }
+
         $grade = Grade::find($slug);
 
-        $grade->grade = request('grade');
-        $grade->description = request('description');
+
+        $gradeValue = request('grade');
+        if ($gradeValue != null) {
+            $grade->grade = $gradeValue;
+        }
+
+        $description = request('description');
+
+        if ($description != null) {
+            $grade->description = $description;
+        }
 
         $grade->save();
 
@@ -47,27 +77,41 @@ class GradeController extends Controller
 
     public function delete($slug)
     {
+        if (auth()->user()->role == 'student') {
+            return redirect()->back();
+        }
+
         $grade = Grade::find($slug);
         $grade->delete();
 
         return redirect('/grades/list')->with('success', 'Stock updated.');
     }
+
     public function add()
     {
+        if (auth()->user()->role == 'student') {
+            return redirect()->back();
+        }
+
         $students = DB::table('users')->where('role', '=', 'student')
             ->get();
+
         return view('add-grade', compact('students'));
     }
 
     public function save()
     {
+        if (auth()->user()->role == 'student') {
+            return redirect()->back();
+        }
+
         $name = request('student');
 
         $id = DB::table('users')->where('name', '=', $name)->first()->id;
         $grade = new Grade;
 
         $grade->grade = request('grade');
-        $grade->teacher = 1;
+        $grade->teacher = auth()->id();
         $grade->description = request('description');
         $grade->subject = request('subjects');
         $grade->student = $id;
